chore(SFT-2753): moved waitForToken to common script loader
