+++ packages/plugin/src/Resources/js/scripts/front-end/captchas/hcaptcha/checkbox.js
@@ -1,1 +1,1 @@
-(()=>{var R={form:{ready:"freeform-ready",reset:"freeform-on-reset",submit:"freeform-on-submit",removeMessages:"freeform-remove-messages",fieldRemoveMessages:"freeform-remove-field-messages",renderSuccess:"freeform-render-success",renderFieldErrors:"freeform-render-field-errors",renderFormErrors:"freeform-render-form-errors",ajaxBeforeSuccess:"freeform-before-ajax-success",ajaxSuccess:"freeform-ajax-success",ajaxError:"freeform-ajax-error",ajaxBeforeSubmit:"freeform-ajax-before-submit",ajaxAfterSubmit:"freeform-ajax-after-submit",afterFailedSubmit:"freeform-after-failed-submit",handleActions:"freeform-handle-actions"},rules:{applied:"freeform-rules-applied"},table:{onAddRow:"freeform-field-table-on-add-row",afterRowAdded:"freeform-field-table-after-row-added",onRemoveRow:"freeform-field-table-on-remove-row",afterRemoveRow:"freeform-field-table-after-remove-row"},dragAndDrop:{renderPreview:"freeform-field-dnd-on-render-preview",renderPreviewRemoveButton:"freeform-field-dnd-on-render-preview-remove-button",renderErrorContainer:"freeform-field-dnd-render-error-container",showGlobalMessage:"freeform-field-dnd-show-global-message",appendErrors:"freeform-field-dnd-append-errors",clearErrors:"freeform-field-dnd-clear-errors",afterErrors:"freeform-field-dnd-after-errors",onChange:"freeform-field-dnd-on-change",onUploadProgress:"freeform-field-dnd-on-upload-progress"},saveAndContinue:{saveFormhandleToken:"freeform-save-form-handle-token"},scripts:{beforeLoad:"freeform-script-before-load",afterLoad:"freeform-script-after-load"},stylesheets:{beforeLoad:"freeform-stylesheet-before-load",afterLoad:"freeform-stylesheet-after-load"}},m=R;var l=(e,r,o,n)=>{k("add",{elements:e,type:r,callback:o,options:n})};var k=(e,{type:r,elements:o,callback:n,options:t})=>{let a=Array.isArray(r)?r:[r],s=Array.isArray(o)?o:[o];Array.from(s).forEach(i=>{a.forEach(d=>{e==="add"?i.addEventListener(d,n,t):i.removeEventListener(d,n,t)})})};var p=null;window.freeform||(window.freeform={});var w;(w=window.freeform)!=null&&w.captchas||(window.freeform.captchas={loaders:new Map,listeners:new WeakSet,loaderPromises:new Map});var F=e=>new Promise((r,o)=>{let n=String(e);if(document.getElementById(n))return r();let t=document.createElement("script"),{trustedTypes:a}=window;if(a!=null&&a.createPolicy){if(!p)try{p=a.createPolicy("freeform#captcha-loader",{createScriptURL:s=>s})}catch{}t.src=p?p.createScriptURL(String(e)):String(e)}else t.src=String(e);t.async=!0,t.defer=!0,t.id=n,t.addEventListener("load",()=>r()),t.addEventListener("error",()=>o(new Error(`Error loading script ${e}`))),document.body.appendChild(t)}),b=(e,r,o,n)=>{let t=v(r,o);if(!t)return;let{listeners:a,loaderPromises:s,loaders:i}=window.freeform.captchas,d=E(t),{lazyLoad:H=!1,version:C="default"}=d,x=H&&!n,c=`${r}-${C}`,f;s.has(c)?f=s.get(c):(f=new Promise((S,A)=>{let P=()=>{F(e).then(S).catch(A)};i.set(c,P)}),s.set(c,f));let y=i.get(c);return x?a.has(o)||(l(o,["input"],y,{once:!0}),a.add(o)):y(),f},v=(e,r)=>r.querySelector(`[data-captcha="${e}"]`),E=e=>({sitekey:e.dataset.sitekey||"",theme:e.dataset.theme||"light",size:e.dataset.size||"normal",version:e.dataset.version,lazyLoad:e.dataset.lazyLoad!==void 0,action:e.dataset.action||"submit",locale:e.dataset.locale});var j="https://js.hcaptcha.com/1/api.js?render=explicit",T="hcaptcha";var h=(e,r)=>{let o=u(e);if(!o)return Promise.resolve();let n=new URL(j),{locale:t}=L(o);return t&&n.searchParams.append("hl",t),b(n,T,e,r)},u=e=>v(T,e),L=e=>E(e);var M,g=e=>{let r=u(e.form);if(!r)return null;let o=r.querySelector(".h-captcha");if(o)return o;let n=document.createElement("div");n.classList.add("h-captcha");let{sitekey:t,theme:a,size:s}=L(r);return r.appendChild(n),M=hcaptcha.render(n,{sitekey:t,theme:a,size:s}),n},z=e=>new Promise(r=>{let o=0,n=setInterval(()=>{let t=e.querySelector('[name="h-captcha-response"]');if(t!=null&&t.value){clearInterval(n),r();return}o+=100,o>=8e3&&(clearInterval(n),r())},100)});document.addEventListener(m.form.submit,e=>{!u(e.form)||e.isBackButtonPressed||e.addCallback(async()=>{try{await h(e.form,!0)}catch{return}g(e),await z(e.form)})});document.addEventListener(m.form.ready,e=>{h(e.form).then(()=>{g(e)}).catch(()=>{})});l(document,[m.form.ajaxAfterSubmit],async e=>{h(e.form).then(()=>{g(e)&&hcaptcha.reset(M)}).catch(()=>{})});})();
+(()=>{var k={form:{ready:"freeform-ready",reset:"freeform-on-reset",submit:"freeform-on-submit",removeMessages:"freeform-remove-messages",fieldRemoveMessages:"freeform-remove-field-messages",renderSuccess:"freeform-render-success",renderFieldErrors:"freeform-render-field-errors",renderFormErrors:"freeform-render-form-errors",ajaxBeforeSuccess:"freeform-before-ajax-success",ajaxSuccess:"freeform-ajax-success",ajaxError:"freeform-ajax-error",ajaxBeforeSubmit:"freeform-ajax-before-submit",ajaxAfterSubmit:"freeform-ajax-after-submit",afterFailedSubmit:"freeform-after-failed-submit",handleActions:"freeform-handle-actions"},rules:{applied:"freeform-rules-applied"},table:{onAddRow:"freeform-field-table-on-add-row",afterRowAdded:"freeform-field-table-after-row-added",onRemoveRow:"freeform-field-table-on-remove-row",afterRemoveRow:"freeform-field-table-after-remove-row"},dragAndDrop:{renderPreview:"freeform-field-dnd-on-render-preview",renderPreviewRemoveButton:"freeform-field-dnd-on-render-preview-remove-button",renderErrorContainer:"freeform-field-dnd-render-error-container",showGlobalMessage:"freeform-field-dnd-show-global-message",appendErrors:"freeform-field-dnd-append-errors",clearErrors:"freeform-field-dnd-clear-errors",afterErrors:"freeform-field-dnd-after-errors",onChange:"freeform-field-dnd-on-change",onUploadProgress:"freeform-field-dnd-on-upload-progress"},saveAndContinue:{saveFormhandleToken:"freeform-save-form-handle-token"},scripts:{beforeLoad:"freeform-script-before-load",afterLoad:"freeform-script-after-load"},stylesheets:{beforeLoad:"freeform-stylesheet-before-load",afterLoad:"freeform-stylesheet-after-load"}},m=k;var l=(e,r,o,n)=>{F("add",{elements:e,type:r,callback:o,options:n})};var F=(e,{type:r,elements:o,callback:n,options:t})=>{let a=Array.isArray(r)?r:[r],s=Array.isArray(o)?o:[o];Array.from(s).forEach(i=>{a.forEach(d=>{e==="add"?i.addEventListener(d,n,t):i.removeEventListener(d,n,t)})})};var p=null;window.freeform||(window.freeform={});var w;(w=window.freeform)!=null&&w.captchas||(window.freeform.captchas={loaders:new Map,listeners:new WeakSet,loaderPromises:new Map});var j=e=>new Promise((r,o)=>{let n=String(e);if(document.getElementById(n))return r();let t=document.createElement("script"),{trustedTypes:a}=window;if(a!=null&&a.createPolicy){if(!p)try{p=a.createPolicy("freeform#captcha-loader",{createScriptURL:s=>s})}catch{}t.src=p?p.createScriptURL(String(e)):String(e)}else t.src=String(e);t.async=!0,t.defer=!0,t.id=n,t.addEventListener("load",()=>r()),t.addEventListener("error",()=>o(new Error(`Error loading script ${e}`))),document.body.appendChild(t)}),T=(e,r,o,n)=>{let t=v(r,o);if(!t)return;let{listeners:a,loaderPromises:s,loaders:i}=window.freeform.captchas,d=E(t),{lazyLoad:C=!1,version:x="default"}=d,S=C&&!n,c=`${r}-${x}`,f;s.has(c)?f=s.get(c):(f=new Promise((A,P)=>{let R=()=>{j(e).then(A).catch(P)};i.set(c,R)}),s.set(c,f));let y=i.get(c);return S?a.has(o)||(l(o,["input"],y,{once:!0}),a.add(o)):y(),f},v=(e,r)=>r.querySelector(`[data-captcha="${e}"]`),E=e=>({sitekey:e.dataset.sitekey||"",theme:e.dataset.theme||"light",size:e.dataset.size||"normal",version:e.dataset.version,lazyLoad:e.dataset.lazyLoad!==void 0,action:e.dataset.action||"submit",locale:e.dataset.locale}),b=(e,r)=>new Promise(o=>{let n=0,t=setInterval(()=>{let a=e.querySelector(`[name="${r}"]`);if(a!=null&&a.value){clearInterval(t),o();return}n+=100,n>=8e3&&(clearInterval(t),o())},100)});var z="https://js.hcaptcha.com/1/api.js?render=explicit",M="hcaptcha";var h=(e,r)=>{let o=u(e);if(!o)return Promise.resolve();let n=new URL(z),{locale:t}=L(o);return t&&n.searchParams.append("hl",t),T(n,M,e,r)},u=e=>v(M,e),L=e=>E(e);var H,g=e=>{let r=u(e.form);if(!r)return null;let o=r.querySelector(".h-captcha");if(o)return o;let n=document.createElement("div");n.classList.add("h-captcha");let{sitekey:t,theme:a,size:s}=L(r);return r.appendChild(n),H=hcaptcha.render(n,{sitekey:t,theme:a,size:s}),n};document.addEventListener(m.form.submit,e=>{!u(e.form)||e.isBackButtonPressed||e.addCallback(async()=>{try{await h(e.form,!0)}catch{return}g(e),await b(e.form,"h-captcha-response")})});document.addEventListener(m.form.ready,e=>{h(e.form).then(()=>{g(e)}).catch(()=>{})});l(document,[m.form.ajaxAfterSubmit],async e=>{h(e.form).then(()=>{g(e)&&hcaptcha.reset(H)}).catch(()=>{})});})();
