+++ packages/plugin/src/Resources/js/scripts/front-end/captchas/hcaptcha/invisible.js
@@ -1,1 +1,1 @@
-(()=>{var S={form:{ready:"freeform-ready",reset:"freeform-on-reset",submit:"freeform-on-submit",removeMessages:"freeform-remove-messages",fieldRemoveMessages:"freeform-remove-field-messages",renderSuccess:"freeform-render-success",renderFieldErrors:"freeform-render-field-errors",renderFormErrors:"freeform-render-form-errors",ajaxBeforeSuccess:"freeform-before-ajax-success",ajaxSuccess:"freeform-ajax-success",ajaxError:"freeform-ajax-error",ajaxBeforeSubmit:"freeform-ajax-before-submit",ajaxAfterSubmit:"freeform-ajax-after-submit",afterFailedSubmit:"freeform-after-failed-submit",handleActions:"freeform-handle-actions"},rules:{applied:"freeform-rules-applied"},table:{onAddRow:"freeform-field-table-on-add-row",afterRowAdded:"freeform-field-table-after-row-added",onRemoveRow:"freeform-field-table-on-remove-row",afterRemoveRow:"freeform-field-table-after-remove-row"},dragAndDrop:{renderPreview:"freeform-field-dnd-on-render-preview",renderPreviewRemoveButton:"freeform-field-dnd-on-render-preview-remove-button",renderErrorContainer:"freeform-field-dnd-render-error-container",showGlobalMessage:"freeform-field-dnd-show-global-message",appendErrors:"freeform-field-dnd-append-errors",clearErrors:"freeform-field-dnd-clear-errors",afterErrors:"freeform-field-dnd-after-errors",onChange:"freeform-field-dnd-on-change",onUploadProgress:"freeform-field-dnd-on-upload-progress"},saveAndContinue:{saveFormhandleToken:"freeform-save-form-handle-token"},scripts:{beforeLoad:"freeform-script-before-load",afterLoad:"freeform-script-after-load"},stylesheets:{beforeLoad:"freeform-stylesheet-before-load",afterLoad:"freeform-stylesheet-after-load"}},E=S;var L=(e,t,r,n)=>{A("add",{elements:e,type:t,callback:r,options:n})};var A=(e,{type:t,elements:r,callback:n,options:o})=>{let a=Array.isArray(t)?t:[t],s=Array.isArray(r)?r:[r];Array.from(s).forEach(i=>{a.forEach(c=>{e==="add"?i.addEventListener(c,n,o):i.removeEventListener(c,n,o)})})};var m=null;window.freeform||(window.freeform={});var g;(g=window.freeform)!=null&&g.captchas||(window.freeform.captchas={loaders:new Map,listeners:new WeakSet,loaderPromises:new Map});var k=e=>new Promise((t,r)=>{let n=String(e);if(document.getElementById(n))return t();let o=document.createElement("script"),{trustedTypes:a}=window;if(a!=null&&a.createPolicy){if(!m)try{m=a.createPolicy("freeform#captcha-loader",{createScriptURL:s=>s})}catch{}o.src=m?m.createScriptURL(String(e)):String(e)}else o.src=String(e);o.async=!0,o.defer=!0,o.id=n,o.addEventListener("load",()=>t()),o.addEventListener("error",()=>r(new Error(`Error loading script ${e}`))),document.body.appendChild(o)}),y=(e,t,r,n)=>{let o=p(t,r);if(!o)return;let{listeners:a,loaderPromises:s,loaders:i}=window.freeform.captchas,c=u(o),{lazyLoad:T=!1,version:M="default"}=c,C=T&&!n,d=`${t}-${M}`,f;s.has(d)?f=s.get(d):(f=new Promise((H,x)=>{let P=()=>{k(e).then(H).catch(x)};i.set(d,P)}),s.set(d,f));let h=i.get(d);return C?a.has(r)||(L(r,["input"],h,{once:!0}),a.add(r)):h(),f},p=(e,t)=>t.querySelector(`[data-captcha="${e}"]`),u=e=>({sitekey:e.dataset.sitekey||"",theme:e.dataset.theme||"light",size:e.dataset.size||"normal",version:e.dataset.version,lazyLoad:e.dataset.lazyLoad!==void 0,action:e.dataset.action||"submit",locale:e.dataset.locale});var R="https://js.hcaptcha.com/1/api.js?render=explicit",b="hcaptcha";var w=(e,t)=>{let r=l(e);if(!r)return Promise.resolve();let n=new URL(R),{locale:o}=v(r);return o&&n.searchParams.append("hl",o),y(n,b,e,t)},l=e=>p(b,e),v=e=>u(e);var j=e=>{let t=l(e.form);if(!t)return null;let r=t.querySelector("[data-hcaptcha]");return r||(r=document.createElement("div"),r.dataset.hcaptcha="",t.appendChild(r)),r};document.addEventListener(E.form.submit,async e=>{e.addCallback(async()=>{let t=j(e);if(!t||e.isBackButtonPressed)return;await w(e.form,!0);let r,n=new Promise(o=>{let a=l(e.form);if(!a)return;let{sitekey:s}=v(a);r=hcaptcha.render(t,{sitekey:s,size:"invisible",callback:i=>{t.querySelector('*[name="h-captcha-response"]').value=i,o()}})});return hcaptcha.execute(r),n})});})();
+(()=>{var S={form:{ready:"freeform-ready",reset:"freeform-on-reset",submit:"freeform-on-submit",removeMessages:"freeform-remove-messages",fieldRemoveMessages:"freeform-remove-field-messages",renderSuccess:"freeform-render-success",renderFieldErrors:"freeform-render-field-errors",renderFormErrors:"freeform-render-form-errors",ajaxBeforeSuccess:"freeform-before-ajax-success",ajaxSuccess:"freeform-ajax-success",ajaxError:"freeform-ajax-error",ajaxBeforeSubmit:"freeform-ajax-before-submit",ajaxAfterSubmit:"freeform-ajax-after-submit",afterFailedSubmit:"freeform-after-failed-submit",handleActions:"freeform-handle-actions"},rules:{applied:"freeform-rules-applied"},table:{onAddRow:"freeform-field-table-on-add-row",afterRowAdded:"freeform-field-table-after-row-added",onRemoveRow:"freeform-field-table-on-remove-row",afterRemoveRow:"freeform-field-table-after-remove-row"},dragAndDrop:{renderPreview:"freeform-field-dnd-on-render-preview",renderPreviewRemoveButton:"freeform-field-dnd-on-render-preview-remove-button",renderErrorContainer:"freeform-field-dnd-render-error-container",showGlobalMessage:"freeform-field-dnd-show-global-message",appendErrors:"freeform-field-dnd-append-errors",clearErrors:"freeform-field-dnd-clear-errors",afterErrors:"freeform-field-dnd-after-errors",onChange:"freeform-field-dnd-on-change",onUploadProgress:"freeform-field-dnd-on-upload-progress"},saveAndContinue:{saveFormhandleToken:"freeform-save-form-handle-token"},scripts:{beforeLoad:"freeform-script-before-load",afterLoad:"freeform-script-after-load"},stylesheets:{beforeLoad:"freeform-stylesheet-before-load",afterLoad:"freeform-stylesheet-after-load"}},E=S;var L=(e,t,r,n)=>{A("add",{elements:e,type:t,callback:r,options:n})};var A=(e,{type:t,elements:r,callback:n,options:o})=>{let a=Array.isArray(t)?t:[t],s=Array.isArray(r)?r:[r];Array.from(s).forEach(i=>{a.forEach(c=>{e==="add"?i.addEventListener(c,n,o):i.removeEventListener(c,n,o)})})};var l=null;window.freeform||(window.freeform={});var g;(g=window.freeform)!=null&&g.captchas||(window.freeform.captchas={loaders:new Map,listeners:new WeakSet,loaderPromises:new Map});var k=e=>new Promise((t,r)=>{let n=String(e);if(document.getElementById(n))return t();let o=document.createElement("script"),{trustedTypes:a}=window;if(a!=null&&a.createPolicy){if(!l)try{l=a.createPolicy("freeform#captcha-loader",{createScriptURL:s=>s})}catch{}o.src=l?l.createScriptURL(String(e)):String(e)}else o.src=String(e);o.async=!0,o.defer=!0,o.id=n,o.addEventListener("load",()=>t()),o.addEventListener("error",()=>r(new Error(`Error loading script ${e}`))),document.body.appendChild(o)}),y=(e,t,r,n)=>{let o=p(t,r);if(!o)return;let{listeners:a,loaderPromises:s,loaders:i}=window.freeform.captchas,c=u(o),{lazyLoad:T=!1,version:M="default"}=c,H=T&&!n,d=`${t}-${M}`,f;s.has(d)?f=s.get(d):(f=new Promise((C,x)=>{let P=()=>{k(e).then(C).catch(x)};i.set(d,P)}),s.set(d,f));let h=i.get(d);return H?a.has(r)||(L(r,["input"],h,{once:!0}),a.add(r)):h(),f},p=(e,t)=>t.querySelector(`[data-captcha="${e}"]`),u=e=>({sitekey:e.dataset.sitekey||"",theme:e.dataset.theme||"light",size:e.dataset.size||"normal",version:e.dataset.version,lazyLoad:e.dataset.lazyLoad!==void 0,action:e.dataset.action||"submit",locale:e.dataset.locale});var R="https://js.hcaptcha.com/1/api.js?render=explicit",b="hcaptcha";var w=(e,t)=>{let r=m(e);if(!r)return Promise.resolve();let n=new URL(R),{locale:o}=v(r);return o&&n.searchParams.append("hl",o),y(n,b,e,t)},m=e=>p(b,e),v=e=>u(e);var F=e=>{let t=m(e.form);if(!t)return null;let r=t.querySelector("[data-hcaptcha]");return r||(r=document.createElement("div"),r.dataset.hcaptcha="",t.appendChild(r)),r};document.addEventListener(E.form.submit,async e=>{e.addCallback(async()=>{let t=F(e);if(!t||e.isBackButtonPressed)return;await w(e.form,!0);let r,n=new Promise(o=>{let a=m(e.form);if(!a)return;let{sitekey:s}=v(a);r=hcaptcha.render(t,{sitekey:s,size:"invisible",callback:i=>{t.querySelector('*[name="h-captcha-response"]').value=i,o()}})});return hcaptcha.execute(r),n})});})();
