+++ packages/plugin/src/Resources/js/scripts/front-end/captchas/hcaptcha/utils/script-loader.js
@@ -1,1 +1,1 @@
-(()=>{var f=(e,n,t,o)=>{C("add",{elements:e,type:n,callback:t,options:o})};var C=(e,{type:n,elements:t,callback:o,options:r})=>{let a=Array.isArray(n)?n:[n],s=Array.isArray(t)?t:[t];Array.from(s).forEach(i=>{a.forEach(d=>{e==="add"?i.addEventListener(d,o,r):i.removeEventListener(d,o,r)})})};var l=null;window.freeform||(window.freeform={});var h;(h=window.freeform)!=null&&h.captchas||(window.freeform.captchas={loaders:new Map,listeners:new WeakSet,loaderPromises:new Map});var P=e=>new Promise((n,t)=>{let o=String(e);if(document.getElementById(o))return n();let r=document.createElement("script"),{trustedTypes:a}=window;if(a!=null&&a.createPolicy){if(!l)try{l=a.createPolicy("freeform#captcha-loader",{createScriptURL:s=>s})}catch{}r.src=l?l.createScriptURL(String(e)):String(e)}else r.src=String(e);r.async=!0,r.defer=!0,r.id=o,r.addEventListener("load",()=>n()),r.addEventListener("error",()=>t(new Error(`Error loading script ${e}`))),document.body.appendChild(r)}),v=(e,n,t,o)=>{let r=m(n,t);if(!r)return;let{listeners:a,loaderPromises:s,loaders:i}=window.freeform.captchas,d=L(r),{lazyLoad:g=!1,version:u="default"}=d,T=g&&!o,c=`${n}-${u}`,p;s.has(c)?p=s.get(c):(p=new Promise((M,H)=>{let w=()=>{P(e).then(M).catch(H)};i.set(c,w)}),s.set(c,p));let E=i.get(c);return T?a.has(t)||(f(t,["input"],E,{once:!0}),a.add(t)):E(),p},m=(e,n)=>n.querySelector(`[data-captcha="${e}"]`),L=e=>({sitekey:e.dataset.sitekey||"",theme:e.dataset.theme||"light",size:e.dataset.size||"normal",version:e.dataset.version,lazyLoad:e.dataset.lazyLoad!==void 0,action:e.dataset.action||"submit",locale:e.dataset.locale});var A="https://js.hcaptcha.com/1/api.js?render=explicit",y="hcaptcha",S=(t=>(t.CHECKBOX="checkbox",t.INVISIBLE="invisible",t))(S||{}),U=(e,n)=>{let t=x(e);if(!t)return Promise.resolve();let o=new URL(A),{locale:r}=b(t);return r&&o.searchParams.append("hl",r),v(o,y,e,n)},x=e=>m(y,e),b=e=>L(e);})();
+(()=>{var v=(e,n,t,o)=>{C("add",{elements:e,type:n,callback:t,options:o})};var C=(e,{type:n,elements:t,callback:o,options:r})=>{let a=Array.isArray(n)?n:[n],s=Array.isArray(t)?t:[t];Array.from(s).forEach(i=>{a.forEach(d=>{e==="add"?i.addEventListener(d,o,r):i.removeEventListener(d,o,r)})})};var p=null;window.freeform||(window.freeform={});var f;(f=window.freeform)!=null&&f.captchas||(window.freeform.captchas={loaders:new Map,listeners:new WeakSet,loaderPromises:new Map});var P=e=>new Promise((n,t)=>{let o=String(e);if(document.getElementById(o))return n();let r=document.createElement("script"),{trustedTypes:a}=window;if(a!=null&&a.createPolicy){if(!p)try{p=a.createPolicy("freeform#captcha-loader",{createScriptURL:s=>s})}catch{}r.src=p?p.createScriptURL(String(e)):String(e)}else r.src=String(e);r.async=!0,r.defer=!0,r.id=o,r.addEventListener("load",()=>n()),r.addEventListener("error",()=>t(new Error(`Error loading script ${e}`))),document.body.appendChild(r)}),h=(e,n,t,o)=>{let r=m(n,t);if(!r)return;let{listeners:a,loaderPromises:s,loaders:i}=window.freeform.captchas,d=L(r),{lazyLoad:g=!1,version:u="default"}=d,T=g&&!o,c=`${n}-${u}`,l;s.has(c)?l=s.get(c):(l=new Promise((M,H)=>{let w=()=>{P(e).then(M).catch(H)};i.set(c,w)}),s.set(c,l));let E=i.get(c);return T?a.has(t)||(v(t,["input"],E,{once:!0}),a.add(t)):E(),l},m=(e,n)=>n.querySelector(`[data-captcha="${e}"]`),L=e=>({sitekey:e.dataset.sitekey||"",theme:e.dataset.theme||"light",size:e.dataset.size||"normal",version:e.dataset.version,lazyLoad:e.dataset.lazyLoad!==void 0,action:e.dataset.action||"submit",locale:e.dataset.locale});var A="https://js.hcaptcha.com/1/api.js?render=explicit",y="hcaptcha",S=(t=>(t.CHECKBOX="checkbox",t.INVISIBLE="invisible",t))(S||{}),I=(e,n)=>{let t=x(e);if(!t)return Promise.resolve();let o=new URL(A),{locale:r}=b(t);return r&&o.searchParams.append("hl",r),h(o,y,e,n)},x=e=>m(y,e),b=e=>L(e);})();
