+++ packages/plugin/src/Resources/js/scripts/front-end/captchas/recaptcha/utils/script-loader.js
@@ -1,1 +1,1 @@
-(()=>{var v=(e,r,n,a)=>{C("add",{elements:e,type:r,callback:n,options:a})};var C=(e,{type:r,elements:n,callback:a,options:t})=>{let o=Array.isArray(r)?r:[r],s=Array.isArray(n)?n:[n];Array.from(s).forEach(i=>{o.forEach(d=>{e==="add"?i.addEventListener(d,a,t):i.removeEventListener(d,a,t)})})};var l=null;window.freeform||(window.freeform={});var f;(f=window.freeform)!=null&&f.captchas||(window.freeform.captchas={loaders:new Map,listeners:new WeakSet,loaderPromises:new Map});var P=e=>new Promise((r,n)=>{let a=String(e);if(document.getElementById(a))return r();let t=document.createElement("script"),{trustedTypes:o}=window;if(o!=null&&o.createPolicy){if(!l)try{l=o.createPolicy("freeform#captcha-loader",{createScriptURL:s=>s})}catch{}t.src=l?l.createScriptURL(String(e)):String(e)}else t.src=String(e);t.async=!0,t.defer=!0,t.id=a,t.addEventListener("load",()=>r()),t.addEventListener("error",()=>n(new Error(`Error loading script ${e}`))),document.body.appendChild(t)}),h=(e,r,n,a)=>{let t=m(r,n);if(!t)return;let{listeners:o,loaderPromises:s,loaders:i}=window.freeform.captchas,d=L(t),{lazyLoad:g=!1,version:u="default"}=d,T=g&&!a,c=`${r}-${u}`,p;s.has(c)?p=s.get(c):(p=new Promise((M,w)=>{let H=()=>{P(e).then(M).catch(w)};i.set(c,H)}),s.set(c,p));let E=i.get(c);return T?o.has(n)||(v(n,["input"],E,{once:!0}),o.add(n)):E(),p},m=(e,r)=>r.querySelector(`[data-captcha="${e}"]`),L=e=>({sitekey:e.dataset.sitekey||"",theme:e.dataset.theme||"light",size:e.dataset.size||"normal",version:e.dataset.version,lazyLoad:e.dataset.lazyLoad!==void 0,action:e.dataset.action||"submit",locale:e.dataset.locale});var A="https://www.google.com/recaptcha/api.js",y="recaptcha",b=(a=>(a.V2_CHECKBOX="v2-checkbox",a.V2_INVISIBLE="v2-invisible",a.V3="v3",a))(b||{}),U=(e,r)=>{let n=S(e);if(!n)return Promise.resolve();let{version:a,sitekey:t,locale:o}=x(n),s=new URL(A);return a==="v3"?s.searchParams.append("render",t):s.searchParams.append("render","explicit"),o&&s.searchParams.append("hl",o),h(s,y,e,r)},S=e=>m(y,e),x=e=>L(e);})();
+(()=>{var E=(e,r,n,o)=>{C("add",{elements:e,type:r,callback:n,options:o})};var C=(e,{type:r,elements:n,callback:o,options:t})=>{let a=Array.isArray(r)?r:[r],s=Array.isArray(n)?n:[n];Array.from(s).forEach(i=>{a.forEach(d=>{e==="add"?i.addEventListener(d,o,t):i.removeEventListener(d,o,t)})})};var p=null;window.freeform||(window.freeform={});var f;(f=window.freeform)!=null&&f.captchas||(window.freeform.captchas={loaders:new Map,listeners:new WeakSet,loaderPromises:new Map});var P=e=>new Promise((r,n)=>{let o=String(e);if(document.getElementById(o))return r();let t=document.createElement("script"),{trustedTypes:a}=window;if(a!=null&&a.createPolicy){if(!p)try{p=a.createPolicy("freeform#captcha-loader",{createScriptURL:s=>s})}catch{}t.src=p?p.createScriptURL(String(e)):String(e)}else t.src=String(e);t.async=!0,t.defer=!0,t.id=o,t.addEventListener("load",()=>r()),t.addEventListener("error",()=>n(new Error(`Error loading script ${e}`))),document.body.appendChild(t)}),h=(e,r,n,o)=>{let t=m(r,n);if(!t)return;let{listeners:a,loaderPromises:s,loaders:i}=window.freeform.captchas,d=L(t),{lazyLoad:g=!1,version:u="default"}=d,T=g&&!o,c=`${r}-${u}`,l;s.has(c)?l=s.get(c):(l=new Promise((M,w)=>{let H=()=>{P(e).then(M).catch(w)};i.set(c,H)}),s.set(c,l));let v=i.get(c);return T?a.has(n)||(E(n,["input"],v,{once:!0}),a.add(n)):v(),l},m=(e,r)=>r.querySelector(`[data-captcha="${e}"]`),L=e=>({sitekey:e.dataset.sitekey||"",theme:e.dataset.theme||"light",size:e.dataset.size||"normal",version:e.dataset.version,lazyLoad:e.dataset.lazyLoad!==void 0,action:e.dataset.action||"submit",locale:e.dataset.locale});var A="https://www.google.com/recaptcha/api.js",y="recaptcha",S=(o=>(o.V2_CHECKBOX="v2-checkbox",o.V2_INVISIBLE="v2-invisible",o.V3="v3",o))(S||{}),I=(e,r)=>{let n=b(e);if(!n)return Promise.resolve();let{version:o,sitekey:t,locale:a}=x(n),s=new URL(A);return o==="v3"?s.searchParams.append("render",t):s.searchParams.append("render","explicit"),a&&s.searchParams.append("hl",a),h(s,y,e,r)},b=e=>m(y,e),x=e=>L(e);})();
