+++ packages/plugin/src/Resources/js/scripts/front-end/captchas/recaptcha/v2-checkbox.js
@@ -1,1 +1,1 @@
-(()=>{var A={form:{ready:"freeform-ready",reset:"freeform-on-reset",submit:"freeform-on-submit",removeMessages:"freeform-remove-messages",fieldRemoveMessages:"freeform-remove-field-messages",renderSuccess:"freeform-render-success",renderFieldErrors:"freeform-render-field-errors",renderFormErrors:"freeform-render-form-errors",ajaxBeforeSuccess:"freeform-before-ajax-success",ajaxSuccess:"freeform-ajax-success",ajaxError:"freeform-ajax-error",ajaxBeforeSubmit:"freeform-ajax-before-submit",ajaxAfterSubmit:"freeform-ajax-after-submit",afterFailedSubmit:"freeform-after-failed-submit",handleActions:"freeform-handle-actions"},rules:{applied:"freeform-rules-applied"},table:{onAddRow:"freeform-field-table-on-add-row",afterRowAdded:"freeform-field-table-after-row-added",onRemoveRow:"freeform-field-table-on-remove-row",afterRemoveRow:"freeform-field-table-after-remove-row"},dragAndDrop:{renderPreview:"freeform-field-dnd-on-render-preview",renderPreviewRemoveButton:"freeform-field-dnd-on-render-preview-remove-button",renderErrorContainer:"freeform-field-dnd-render-error-container",showGlobalMessage:"freeform-field-dnd-show-global-message",appendErrors:"freeform-field-dnd-append-errors",clearErrors:"freeform-field-dnd-clear-errors",afterErrors:"freeform-field-dnd-after-errors",onChange:"freeform-field-dnd-on-change",onUploadProgress:"freeform-field-dnd-on-upload-progress"},saveAndContinue:{saveFormhandleToken:"freeform-save-form-handle-token"},scripts:{beforeLoad:"freeform-script-before-load",afterLoad:"freeform-script-after-load"},stylesheets:{beforeLoad:"freeform-stylesheet-before-load",afterLoad:"freeform-stylesheet-after-load"}},m=A;var l=(e,t,r,a)=>{R("add",{elements:e,type:t,callback:r,options:a})};var R=(e,{type:t,elements:r,callback:a,options:o})=>{let n=Array.isArray(t)?t:[t],s=Array.isArray(r)?r:[r];Array.from(s).forEach(i=>{n.forEach(c=>{e==="add"?i.addEventListener(c,a,o):i.removeEventListener(c,a,o)})})};var p=null;window.freeform||(window.freeform={});var w;(w=window.freeform)!=null&&w.captchas||(window.freeform.captchas={loaders:new Map,listeners:new WeakSet,loaderPromises:new Map});var k=e=>new Promise((t,r)=>{let a=String(e);if(document.getElementById(a))return t();let o=document.createElement("script"),{trustedTypes:n}=window;if(n!=null&&n.createPolicy){if(!p)try{p=n.createPolicy("freeform#captcha-loader",{createScriptURL:s=>s})}catch{}o.src=p?p.createScriptURL(String(e)):String(e)}else o.src=String(e);o.async=!0,o.defer=!0,o.id=a,o.addEventListener("load",()=>t()),o.addEventListener("error",()=>r(new Error(`Error loading script ${e}`))),document.body.appendChild(o)}),b=(e,t,r,a)=>{let o=v(t,r);if(!o)return;let{listeners:n,loaderPromises:s,loaders:i}=window.freeform.captchas,c=E(o),{lazyLoad:M=!1,version:C="default"}=c,H=M&&!a,d=`${t}-${C}`,f;s.has(d)?f=s.get(d):(f=new Promise((x,P)=>{let S=()=>{k(e).then(x).catch(P)};i.set(d,S)}),s.set(d,f));let y=i.get(d);return H?n.has(r)||(l(r,["input"],y,{once:!0}),n.add(r)):y(),f},v=(e,t)=>t.querySelector(`[data-captcha="${e}"]`),E=e=>({sitekey:e.dataset.sitekey||"",theme:e.dataset.theme||"light",size:e.dataset.size||"normal",version:e.dataset.version,lazyLoad:e.dataset.lazyLoad!==void 0,action:e.dataset.action||"submit",locale:e.dataset.locale});var F="https://www.google.com/recaptcha/api.js",T="recaptcha";var u=(e,t)=>{let r=h(e);if(!r)return Promise.resolve();let{version:a,sitekey:o,locale:n}=L(r),s=new URL(F);return a==="v3"?s.searchParams.append("render",o):s.searchParams.append("render","explicit"),n&&s.searchParams.append("hl",n),b(s,T,e,t)},h=e=>v(T,e),L=e=>E(e);var g=e=>{let t=h(e.form);if(!t)return null;let r=e.form.querySelector(".g-recaptcha");if(r)return r;r=document.createElement("div"),r.classList.add("g-recaptcha");let{sitekey:a,theme:o,size:n}=L(t);return t.appendChild(r),grecaptcha.ready(()=>{let s=grecaptcha.render(r,{sitekey:a,theme:o,size:n});r.dataset.captchaId=String(s)}),r},j=e=>new Promise(t=>{let r=0,a=setInterval(()=>{let o=e.querySelector('[name="g-recaptcha-response"]');if(o!=null&&o.value){clearInterval(a),t();return}r+=100,r>=8e3&&(clearInterval(a),t())},100)});document.addEventListener(m.form.submit,e=>{!h(e.form)||e.isBackButtonPressed||e.addCallback(async()=>{try{await u(e.form,!0)}catch{return}g(e),await j(e.form)})});document.addEventListener(m.form.ready,e=>{u(e.form).then(()=>{g(e)}).catch(()=>{})});l(document,[m.form.ajaxAfterSubmit],async e=>{u(e.form,!0).then(()=>{let t=g(e);if(t){let r=t.dataset.captchaId;grecaptcha.ready(()=>grecaptcha.reset(r?Number(r):void 0))}}).catch(()=>{})});})();
+(()=>{var k={form:{ready:"freeform-ready",reset:"freeform-on-reset",submit:"freeform-on-submit",removeMessages:"freeform-remove-messages",fieldRemoveMessages:"freeform-remove-field-messages",renderSuccess:"freeform-render-success",renderFieldErrors:"freeform-render-field-errors",renderFormErrors:"freeform-render-form-errors",ajaxBeforeSuccess:"freeform-before-ajax-success",ajaxSuccess:"freeform-ajax-success",ajaxError:"freeform-ajax-error",ajaxBeforeSubmit:"freeform-ajax-before-submit",ajaxAfterSubmit:"freeform-ajax-after-submit",afterFailedSubmit:"freeform-after-failed-submit",handleActions:"freeform-handle-actions"},rules:{applied:"freeform-rules-applied"},table:{onAddRow:"freeform-field-table-on-add-row",afterRowAdded:"freeform-field-table-after-row-added",onRemoveRow:"freeform-field-table-on-remove-row",afterRemoveRow:"freeform-field-table-after-remove-row"},dragAndDrop:{renderPreview:"freeform-field-dnd-on-render-preview",renderPreviewRemoveButton:"freeform-field-dnd-on-render-preview-remove-button",renderErrorContainer:"freeform-field-dnd-render-error-container",showGlobalMessage:"freeform-field-dnd-show-global-message",appendErrors:"freeform-field-dnd-append-errors",clearErrors:"freeform-field-dnd-clear-errors",afterErrors:"freeform-field-dnd-after-errors",onChange:"freeform-field-dnd-on-change",onUploadProgress:"freeform-field-dnd-on-upload-progress"},saveAndContinue:{saveFormhandleToken:"freeform-save-form-handle-token"},scripts:{beforeLoad:"freeform-script-before-load",afterLoad:"freeform-script-after-load"},stylesheets:{beforeLoad:"freeform-stylesheet-before-load",afterLoad:"freeform-stylesheet-after-load"}},m=k;var l=(e,t,r,n)=>{R("add",{elements:e,type:t,callback:r,options:n})};var R=(e,{type:t,elements:r,callback:n,options:o})=>{let a=Array.isArray(t)?t:[t],s=Array.isArray(r)?r:[r];Array.from(s).forEach(i=>{a.forEach(c=>{e==="add"?i.addEventListener(c,n,o):i.removeEventListener(c,n,o)})})};var p=null;window.freeform||(window.freeform={});var w;(w=window.freeform)!=null&&w.captchas||(window.freeform.captchas={loaders:new Map,listeners:new WeakSet,loaderPromises:new Map});var F=e=>new Promise((t,r)=>{let n=String(e);if(document.getElementById(n))return t();let o=document.createElement("script"),{trustedTypes:a}=window;if(a!=null&&a.createPolicy){if(!p)try{p=a.createPolicy("freeform#captcha-loader",{createScriptURL:s=>s})}catch{}o.src=p?p.createScriptURL(String(e)):String(e)}else o.src=String(e);o.async=!0,o.defer=!0,o.id=n,o.addEventListener("load",()=>t()),o.addEventListener("error",()=>r(new Error(`Error loading script ${e}`))),document.body.appendChild(o)}),b=(e,t,r,n)=>{let o=v(t,r);if(!o)return;let{listeners:a,loaderPromises:s,loaders:i}=window.freeform.captchas,c=E(o),{lazyLoad:C=!1,version:H="default"}=c,x=C&&!n,d=`${t}-${H}`,f;s.has(d)?f=s.get(d):(f=new Promise((P,S)=>{let A=()=>{F(e).then(P).catch(S)};i.set(d,A)}),s.set(d,f));let y=i.get(d);return x?a.has(r)||(l(r,["input"],y,{once:!0}),a.add(r)):y(),f},v=(e,t)=>t.querySelector(`[data-captcha="${e}"]`),E=e=>({sitekey:e.dataset.sitekey||"",theme:e.dataset.theme||"light",size:e.dataset.size||"normal",version:e.dataset.version,lazyLoad:e.dataset.lazyLoad!==void 0,action:e.dataset.action||"submit",locale:e.dataset.locale}),T=(e,t)=>new Promise(r=>{let n=0,o=setInterval(()=>{let a=e.querySelector(`[name="${t}"]`);if(a!=null&&a.value){clearInterval(o),r();return}n+=100,n>=8e3&&(clearInterval(o),r())},100)});var j="https://www.google.com/recaptcha/api.js",M="recaptcha";var u=(e,t)=>{let r=h(e);if(!r)return Promise.resolve();let{version:n,sitekey:o,locale:a}=L(r),s=new URL(j);return n==="v3"?s.searchParams.append("render",o):s.searchParams.append("render","explicit"),a&&s.searchParams.append("hl",a),b(s,M,e,t)},h=e=>v(M,e),L=e=>E(e);var g=e=>{let t=h(e.form);if(!t)return null;let r=e.form.querySelector(".g-recaptcha");if(r)return r;r=document.createElement("div"),r.classList.add("g-recaptcha");let{sitekey:n,theme:o,size:a}=L(t);return t.appendChild(r),grecaptcha.ready(()=>{let s=grecaptcha.render(r,{sitekey:n,theme:o,size:a});r.dataset.captchaId=String(s)}),r};document.addEventListener(m.form.submit,e=>{!h(e.form)||e.isBackButtonPressed||e.addCallback(async()=>{try{await u(e.form,!0)}catch{return}g(e),await T(e.form,"g-recaptcha-response")})});document.addEventListener(m.form.ready,e=>{u(e.form).then(()=>{g(e)}).catch(()=>{})});l(document,[m.form.ajaxAfterSubmit],async e=>{u(e.form,!0).then(()=>{let t=g(e);if(t){let r=t.dataset.captchaId;grecaptcha.ready(()=>grecaptcha.reset(r?Number(r):void 0))}}).catch(()=>{})});})();
