+++ packages/plugin/src/Resources/js/scripts/front-end/captchas/recaptcha/v2-invisible.js
@@ -1,1 +1,1 @@
-(()=>{var k={form:{ready:"freeform-ready",reset:"freeform-on-reset",submit:"freeform-on-submit",removeMessages:"freeform-remove-messages",fieldRemoveMessages:"freeform-remove-field-messages",renderSuccess:"freeform-render-success",renderFieldErrors:"freeform-render-field-errors",renderFormErrors:"freeform-render-form-errors",ajaxBeforeSuccess:"freeform-before-ajax-success",ajaxSuccess:"freeform-ajax-success",ajaxError:"freeform-ajax-error",ajaxBeforeSubmit:"freeform-ajax-before-submit",ajaxAfterSubmit:"freeform-ajax-after-submit",afterFailedSubmit:"freeform-after-failed-submit",handleActions:"freeform-handle-actions"},rules:{applied:"freeform-rules-applied"},table:{onAddRow:"freeform-field-table-on-add-row",afterRowAdded:"freeform-field-table-after-row-added",onRemoveRow:"freeform-field-table-on-remove-row",afterRemoveRow:"freeform-field-table-after-remove-row"},dragAndDrop:{renderPreview:"freeform-field-dnd-on-render-preview",renderPreviewRemoveButton:"freeform-field-dnd-on-render-preview-remove-button",renderErrorContainer:"freeform-field-dnd-render-error-container",showGlobalMessage:"freeform-field-dnd-show-global-message",appendErrors:"freeform-field-dnd-append-errors",clearErrors:"freeform-field-dnd-clear-errors",afterErrors:"freeform-field-dnd-after-errors",onChange:"freeform-field-dnd-on-change",onUploadProgress:"freeform-field-dnd-on-upload-progress"},saveAndContinue:{saveFormhandleToken:"freeform-save-form-handle-token"},scripts:{beforeLoad:"freeform-script-before-load",afterLoad:"freeform-script-after-load"},stylesheets:{beforeLoad:"freeform-stylesheet-before-load",afterLoad:"freeform-stylesheet-after-load"}},d=k;var l=(e,t,r,o)=>{R("add",{elements:e,type:t,callback:r,options:o})};var R=(e,{type:t,elements:r,callback:o,options:a})=>{let n=Array.isArray(t)?t:[t],s=Array.isArray(r)?r:[r];Array.from(s).forEach(i=>{n.forEach(f=>{e==="add"?i.addEventListener(f,o,a):i.removeEventListener(f,o,a)})})};var p=null;window.freeform||(window.freeform={});var y;(y=window.freeform)!=null&&y.captchas||(window.freeform.captchas={loaders:new Map,listeners:new WeakSet,loaderPromises:new Map});var F=e=>new Promise((t,r)=>{let o=String(e);if(document.getElementById(o))return t();let a=document.createElement("script"),{trustedTypes:n}=window;if(n!=null&&n.createPolicy){if(!p)try{p=n.createPolicy("freeform#captcha-loader",{createScriptURL:s=>s})}catch{}a.src=p?p.createScriptURL(String(e)):String(e)}else a.src=String(e);a.async=!0,a.defer=!0,a.id=o,a.addEventListener("load",()=>t()),a.addEventListener("error",()=>r(new Error(`Error loading script ${e}`))),document.body.appendChild(a)}),b=(e,t,r,o)=>{let a=v(t,r);if(!a)return;let{listeners:n,loaderPromises:s,loaders:i}=window.freeform.captchas,f=h(a),{lazyLoad:C=!1,version:H="default"}=f,x=C&&!o,c=`${t}-${H}`,m;s.has(c)?m=s.get(c):(m=new Promise((S,P)=>{let A=()=>{F(e).then(S).catch(P)};i.set(c,A)}),s.set(c,m));let g=i.get(c);return x?n.has(r)||(l(r,["input"],g,{once:!0}),n.add(r)):g(),m},v=(e,t)=>t.querySelector(`[data-captcha="${e}"]`),h=e=>({sitekey:e.dataset.sitekey||"",theme:e.dataset.theme||"light",size:e.dataset.size||"normal",version:e.dataset.version,lazyLoad:e.dataset.lazyLoad!==void 0,action:e.dataset.action||"submit",locale:e.dataset.locale});var j="https://www.google.com/recaptcha/api.js",w="recaptcha";var E=(e,t)=>{let r=u(e);if(!r)return Promise.resolve();let{version:o,sitekey:a,locale:n}=L(r),s=new URL(j);return o==="v3"?s.searchParams.append("render",a):s.searchParams.append("render","explicit"),n&&s.searchParams.append("hl",n),b(s,w,e,t)},u=e=>v(w,e),L=e=>h(e);var T,M=e=>{let t=u(e.form);if(!t)return null;let r=t.querySelector("[data-recaptcha]");return r||(r=document.createElement("div"),r.dataset.recaptcha="",t.appendChild(r)),r},z=e=>{E(e.form).then(()=>{let t=u(e.form);if(!t)return;let{sitekey:r}=L(t),o=M(e);o&&(o.innerHTML?grecaptcha.ready(grecaptcha.reset):grecaptcha.ready(()=>{let a=grecaptcha.render(o,{sitekey:r,size:"invisible",callback:n=>{o.querySelector('*[name="g-recaptcha-response"]').value=n,T()}});o.dataset.captchaId=String(a)}))}).catch(()=>{})};document.addEventListener(d.form.submit,async e=>{e.addCallback(async()=>{let t=new Promise(o=>{T=o}),r=M(e);if(!(!r||e.isBackButtonPressed)){try{await E(e.form,!0)}catch{return}return grecaptcha.ready(()=>{let o=r.dataset.captchaId;grecaptcha.execute(o&&Number(o))}),t}})});l(document,[d.form.ready,d.form.afterFailedSubmit,d.form.ajaxAfterSubmit],z);})();
+(()=>{var k={form:{ready:"freeform-ready",reset:"freeform-on-reset",submit:"freeform-on-submit",removeMessages:"freeform-remove-messages",fieldRemoveMessages:"freeform-remove-field-messages",renderSuccess:"freeform-render-success",renderFieldErrors:"freeform-render-field-errors",renderFormErrors:"freeform-render-form-errors",ajaxBeforeSuccess:"freeform-before-ajax-success",ajaxSuccess:"freeform-ajax-success",ajaxError:"freeform-ajax-error",ajaxBeforeSubmit:"freeform-ajax-before-submit",ajaxAfterSubmit:"freeform-ajax-after-submit",afterFailedSubmit:"freeform-after-failed-submit",handleActions:"freeform-handle-actions"},rules:{applied:"freeform-rules-applied"},table:{onAddRow:"freeform-field-table-on-add-row",afterRowAdded:"freeform-field-table-after-row-added",onRemoveRow:"freeform-field-table-on-remove-row",afterRemoveRow:"freeform-field-table-after-remove-row"},dragAndDrop:{renderPreview:"freeform-field-dnd-on-render-preview",renderPreviewRemoveButton:"freeform-field-dnd-on-render-preview-remove-button",renderErrorContainer:"freeform-field-dnd-render-error-container",showGlobalMessage:"freeform-field-dnd-show-global-message",appendErrors:"freeform-field-dnd-append-errors",clearErrors:"freeform-field-dnd-clear-errors",afterErrors:"freeform-field-dnd-after-errors",onChange:"freeform-field-dnd-on-change",onUploadProgress:"freeform-field-dnd-on-upload-progress"},saveAndContinue:{saveFormhandleToken:"freeform-save-form-handle-token"},scripts:{beforeLoad:"freeform-script-before-load",afterLoad:"freeform-script-after-load"},stylesheets:{beforeLoad:"freeform-stylesheet-before-load",afterLoad:"freeform-stylesheet-after-load"}},d=k;var l=(e,t,r,o)=>{R("add",{elements:e,type:t,callback:r,options:o})};var R=(e,{type:t,elements:r,callback:o,options:a})=>{let n=Array.isArray(t)?t:[t],s=Array.isArray(r)?r:[r];Array.from(s).forEach(i=>{n.forEach(f=>{e==="add"?i.addEventListener(f,o,a):i.removeEventListener(f,o,a)})})};var p=null;window.freeform||(window.freeform={});var y;(y=window.freeform)!=null&&y.captchas||(window.freeform.captchas={loaders:new Map,listeners:new WeakSet,loaderPromises:new Map});var F=e=>new Promise((t,r)=>{let o=String(e);if(document.getElementById(o))return t();let a=document.createElement("script"),{trustedTypes:n}=window;if(n!=null&&n.createPolicy){if(!p)try{p=n.createPolicy("freeform#captcha-loader",{createScriptURL:s=>s})}catch{}a.src=p?p.createScriptURL(String(e)):String(e)}else a.src=String(e);a.async=!0,a.defer=!0,a.id=o,a.addEventListener("load",()=>t()),a.addEventListener("error",()=>r(new Error(`Error loading script ${e}`))),document.body.appendChild(a)}),b=(e,t,r,o)=>{let a=v(t,r);if(!a)return;let{listeners:n,loaderPromises:s,loaders:i}=window.freeform.captchas,f=h(a),{lazyLoad:H=!1,version:C="default"}=f,x=H&&!o,c=`${t}-${C}`,m;s.has(c)?m=s.get(c):(m=new Promise((P,S)=>{let A=()=>{F(e).then(P).catch(S)};i.set(c,A)}),s.set(c,m));let g=i.get(c);return x?n.has(r)||(l(r,["input"],g,{once:!0}),n.add(r)):g(),m},v=(e,t)=>t.querySelector(`[data-captcha="${e}"]`),h=e=>({sitekey:e.dataset.sitekey||"",theme:e.dataset.theme||"light",size:e.dataset.size||"normal",version:e.dataset.version,lazyLoad:e.dataset.lazyLoad!==void 0,action:e.dataset.action||"submit",locale:e.dataset.locale});var j="https://www.google.com/recaptcha/api.js",w="recaptcha";var E=(e,t)=>{let r=u(e);if(!r)return Promise.resolve();let{version:o,sitekey:a,locale:n}=L(r),s=new URL(j);return o==="v3"?s.searchParams.append("render",a):s.searchParams.append("render","explicit"),n&&s.searchParams.append("hl",n),b(s,w,e,t)},u=e=>v(w,e),L=e=>h(e);var T,M=e=>{let t=u(e.form);if(!t)return null;let r=t.querySelector("[data-recaptcha]");return r||(r=document.createElement("div"),r.dataset.recaptcha="",t.appendChild(r)),r},I=e=>{E(e.form).then(()=>{let t=u(e.form);if(!t)return;let{sitekey:r}=L(t),o=M(e);o&&(o.innerHTML?grecaptcha.ready(grecaptcha.reset):grecaptcha.ready(()=>{let a=grecaptcha.render(o,{sitekey:r,size:"invisible",callback:n=>{o.querySelector('*[name="g-recaptcha-response"]').value=n,T()}});o.dataset.captchaId=String(a)}))}).catch(()=>{})};document.addEventListener(d.form.submit,async e=>{e.addCallback(async()=>{let t=new Promise(o=>{T=o}),r=M(e);if(!(!r||e.isBackButtonPressed)){try{await E(e.form,!0)}catch{return}return grecaptcha.ready(()=>{let o=r.dataset.captchaId;grecaptcha.execute(o&&Number(o))}),t}})});l(document,[d.form.ready,d.form.afterFailedSubmit,d.form.ajaxAfterSubmit],I);})();
