+++ packages/plugin/src/Resources/js/scripts/front-end/captchas/recaptcha/v3.js
@@ -1,1 +1,1 @@
-(()=>{var P={form:{ready:"freeform-ready",reset:"freeform-on-reset",submit:"freeform-on-submit",removeMessages:"freeform-remove-messages",fieldRemoveMessages:"freeform-remove-field-messages",renderSuccess:"freeform-render-success",renderFieldErrors:"freeform-render-field-errors",renderFormErrors:"freeform-render-form-errors",ajaxBeforeSuccess:"freeform-before-ajax-success",ajaxSuccess:"freeform-ajax-success",ajaxError:"freeform-ajax-error",ajaxBeforeSubmit:"freeform-ajax-before-submit",ajaxAfterSubmit:"freeform-ajax-after-submit",afterFailedSubmit:"freeform-after-failed-submit",handleActions:"freeform-handle-actions"},rules:{applied:"freeform-rules-applied"},table:{onAddRow:"freeform-field-table-on-add-row",afterRowAdded:"freeform-field-table-after-row-added",onRemoveRow:"freeform-field-table-on-remove-row",afterRemoveRow:"freeform-field-table-after-remove-row"},dragAndDrop:{renderPreview:"freeform-field-dnd-on-render-preview",renderPreviewRemoveButton:"freeform-field-dnd-on-render-preview-remove-button",renderErrorContainer:"freeform-field-dnd-render-error-container",showGlobalMessage:"freeform-field-dnd-show-global-message",appendErrors:"freeform-field-dnd-append-errors",clearErrors:"freeform-field-dnd-clear-errors",afterErrors:"freeform-field-dnd-after-errors",onChange:"freeform-field-dnd-on-change",onUploadProgress:"freeform-field-dnd-on-upload-progress"},saveAndContinue:{saveFormhandleToken:"freeform-save-form-handle-token"},scripts:{beforeLoad:"freeform-script-before-load",afterLoad:"freeform-script-after-load"},stylesheets:{beforeLoad:"freeform-stylesheet-before-load",afterLoad:"freeform-stylesheet-after-load"}},p=P;var L=(e,t,r,s)=>{S("add",{elements:e,type:t,callback:r,options:s})};var S=(e,{type:t,elements:r,callback:s,options:o})=>{let a=Array.isArray(t)?t:[t],n=Array.isArray(r)?r:[r];Array.from(n).forEach(i=>{a.forEach(c=>{e==="add"?i.addEventListener(c,s,o):i.removeEventListener(c,s,o)})})};var l=null;window.freeform||(window.freeform={});var g;(g=window.freeform)!=null&&g.captchas||(window.freeform.captchas={loaders:new Map,listeners:new WeakSet,loaderPromises:new Map});var k=e=>new Promise((t,r)=>{let s=String(e);if(document.getElementById(s))return t();let o=document.createElement("script"),{trustedTypes:a}=window;if(a!=null&&a.createPolicy){if(!l)try{l=a.createPolicy("freeform#captcha-loader",{createScriptURL:n=>n})}catch{}o.src=l?l.createScriptURL(String(e)):String(e)}else o.src=String(e);o.async=!0,o.defer=!0,o.id=s,o.addEventListener("load",()=>t()),o.addEventListener("error",()=>r(new Error(`Error loading script ${e}`))),document.body.appendChild(o)}),w=(e,t,r,s)=>{let o=u(t,r);if(!o)return;let{listeners:a,loaderPromises:n,loaders:i}=window.freeform.captchas,c=v(o),{lazyLoad:T=!1,version:x="default"}=c,M=T&&!s,d=`${t}-${x}`,f;n.has(d)?f=n.get(d):(f=new Promise((C,H)=>{let A=()=>{k(e).then(C).catch(H)};i.set(d,A)}),n.set(d,f));let y=i.get(d);return M?a.has(r)||(L(r,["input"],y,{once:!0}),a.add(r)):y(),f},u=(e,t)=>t.querySelector(`[data-captcha="${e}"]`),v=e=>({sitekey:e.dataset.sitekey||"",theme:e.dataset.theme||"light",size:e.dataset.size||"normal",version:e.dataset.version,lazyLoad:e.dataset.lazyLoad!==void 0,action:e.dataset.action||"submit",locale:e.dataset.locale});var R="https://www.google.com/recaptcha/api.js",b="recaptcha";var h=(e,t)=>{let r=m(e);if(!r)return Promise.resolve();let{version:s,sitekey:o,locale:a}=E(r),n=new URL(R);return s==="v3"?n.searchParams.append("render",o):n.searchParams.append("render","explicit"),a&&n.searchParams.append("hl",a),w(n,b,e,t)},m=e=>u(b,e),E=e=>v(e);var F=e=>{let t=m(e.form);if(!t)return null;let r=t.querySelector("[data-recaptcha]");return r||(r=document.createElement("textarea"),r.dataset.recaptcha="",r.name="g-recaptcha-response",r.style.visibility="hidden",r.style.position="absolute",r.style.top="-9999px",r.style.left="-9999px",r.style.width="1px",r.style.height="1px",r.style.overflow="hidden",r.style.border="none",t.appendChild(r)),r};document.addEventListener(p.form.ready,e=>{h(e.form).catch(()=>{})});document.addEventListener(p.form.submit,e=>{e.addCallback(async()=>{let t=m(e.form);if(!t)return null;let r=F(e);if(!r||e.isBackButtonPressed)return;try{await h(e.form,!0)}catch{return}let{sitekey:s,action:o}=E(t);return await new Promise(a=>{grecaptcha.ready(async()=>{let n=await grecaptcha.execute(s,{action:o});r.value=n,a()})})})});})();
+(()=>{var A={form:{ready:"freeform-ready",reset:"freeform-on-reset",submit:"freeform-on-submit",removeMessages:"freeform-remove-messages",fieldRemoveMessages:"freeform-remove-field-messages",renderSuccess:"freeform-render-success",renderFieldErrors:"freeform-render-field-errors",renderFormErrors:"freeform-render-form-errors",ajaxBeforeSuccess:"freeform-before-ajax-success",ajaxSuccess:"freeform-ajax-success",ajaxError:"freeform-ajax-error",ajaxBeforeSubmit:"freeform-ajax-before-submit",ajaxAfterSubmit:"freeform-ajax-after-submit",afterFailedSubmit:"freeform-after-failed-submit",handleActions:"freeform-handle-actions"},rules:{applied:"freeform-rules-applied"},table:{onAddRow:"freeform-field-table-on-add-row",afterRowAdded:"freeform-field-table-after-row-added",onRemoveRow:"freeform-field-table-on-remove-row",afterRemoveRow:"freeform-field-table-after-remove-row"},dragAndDrop:{renderPreview:"freeform-field-dnd-on-render-preview",renderPreviewRemoveButton:"freeform-field-dnd-on-render-preview-remove-button",renderErrorContainer:"freeform-field-dnd-render-error-container",showGlobalMessage:"freeform-field-dnd-show-global-message",appendErrors:"freeform-field-dnd-append-errors",clearErrors:"freeform-field-dnd-clear-errors",afterErrors:"freeform-field-dnd-after-errors",onChange:"freeform-field-dnd-on-change",onUploadProgress:"freeform-field-dnd-on-upload-progress"},saveAndContinue:{saveFormhandleToken:"freeform-save-form-handle-token"},scripts:{beforeLoad:"freeform-script-before-load",afterLoad:"freeform-script-after-load"},stylesheets:{beforeLoad:"freeform-stylesheet-before-load",afterLoad:"freeform-stylesheet-after-load"}},p=A;var L=(e,t,r,s)=>{S("add",{elements:e,type:t,callback:r,options:s})};var S=(e,{type:t,elements:r,callback:s,options:o})=>{let a=Array.isArray(t)?t:[t],n=Array.isArray(r)?r:[r];Array.from(n).forEach(i=>{a.forEach(c=>{e==="add"?i.addEventListener(c,s,o):i.removeEventListener(c,s,o)})})};var l=null;window.freeform||(window.freeform={});var g;(g=window.freeform)!=null&&g.captchas||(window.freeform.captchas={loaders:new Map,listeners:new WeakSet,loaderPromises:new Map});var k=e=>new Promise((t,r)=>{let s=String(e);if(document.getElementById(s))return t();let o=document.createElement("script"),{trustedTypes:a}=window;if(a!=null&&a.createPolicy){if(!l)try{l=a.createPolicy("freeform#captcha-loader",{createScriptURL:n=>n})}catch{}o.src=l?l.createScriptURL(String(e)):String(e)}else o.src=String(e);o.async=!0,o.defer=!0,o.id=s,o.addEventListener("load",()=>t()),o.addEventListener("error",()=>r(new Error(`Error loading script ${e}`))),document.body.appendChild(o)}),w=(e,t,r,s)=>{let o=v(t,r);if(!o)return;let{listeners:a,loaderPromises:n,loaders:i}=window.freeform.captchas,c=u(o),{lazyLoad:T=!1,version:M="default"}=c,x=T&&!s,d=`${t}-${M}`,f;n.has(d)?f=n.get(d):(f=new Promise((C,H)=>{let P=()=>{k(e).then(C).catch(H)};i.set(d,P)}),n.set(d,f));let y=i.get(d);return x?a.has(r)||(L(r,["input"],y,{once:!0}),a.add(r)):y(),f},v=(e,t)=>t.querySelector(`[data-captcha="${e}"]`),u=e=>({sitekey:e.dataset.sitekey||"",theme:e.dataset.theme||"light",size:e.dataset.size||"normal",version:e.dataset.version,lazyLoad:e.dataset.lazyLoad!==void 0,action:e.dataset.action||"submit",locale:e.dataset.locale});var R="https://www.google.com/recaptcha/api.js",b="recaptcha";var h=(e,t)=>{let r=m(e);if(!r)return Promise.resolve();let{version:s,sitekey:o,locale:a}=E(r),n=new URL(R);return s==="v3"?n.searchParams.append("render",o):n.searchParams.append("render","explicit"),a&&n.searchParams.append("hl",a),w(n,b,e,t)},m=e=>v(b,e),E=e=>u(e);var F=e=>{let t=m(e.form);if(!t)return null;let r=t.querySelector("[data-recaptcha]");return r||(r=document.createElement("textarea"),r.dataset.recaptcha="",r.name="g-recaptcha-response",r.style.visibility="hidden",r.style.position="absolute",r.style.top="-9999px",r.style.left="-9999px",r.style.width="1px",r.style.height="1px",r.style.overflow="hidden",r.style.border="none",t.appendChild(r)),r};document.addEventListener(p.form.ready,e=>{h(e.form).catch(()=>{})});document.addEventListener(p.form.submit,e=>{e.addCallback(async()=>{let t=m(e.form);if(!t)return null;let r=F(e);if(!r||e.isBackButtonPressed)return;try{await h(e.form,!0)}catch{return}let{sitekey:s,action:o}=E(t);return await new Promise(a=>{grecaptcha.ready(async()=>{let n=await grecaptcha.execute(s,{action:o});r.value=n,a()})})})});})();
