+++ packages/plugin/src/Resources/js/scripts/front-end/captchas/turnstile/utils/script-loader.js
@@ -1,1 +1,1 @@
-(()=>{var v=(e,t,n,o)=>{C("add",{elements:e,type:t,callback:n,options:o})};var C=(e,{type:t,elements:n,callback:o,options:r})=>{let a=Array.isArray(t)?t:[t],s=Array.isArray(n)?n:[n];Array.from(s).forEach(i=>{a.forEach(d=>{e==="add"?i.addEventListener(d,o,r):i.removeEventListener(d,o,r)})})};var p=null;window.freeform||(window.freeform={});var f;(f=window.freeform)!=null&&f.captchas||(window.freeform.captchas={loaders:new Map,listeners:new WeakSet,loaderPromises:new Map});var P=e=>new Promise((t,n)=>{let o=String(e);if(document.getElementById(o))return t();let r=document.createElement("script"),{trustedTypes:a}=window;if(a!=null&&a.createPolicy){if(!p)try{p=a.createPolicy("freeform#captcha-loader",{createScriptURL:s=>s})}catch{}r.src=p?p.createScriptURL(String(e)):String(e)}else r.src=String(e);r.async=!0,r.defer=!0,r.id=o,r.addEventListener("load",()=>t()),r.addEventListener("error",()=>n(new Error(`Error loading script ${e}`))),document.body.appendChild(r)}),h=(e,t,n,o)=>{let r=m(t,n);if(!r)return;let{listeners:a,loaderPromises:s,loaders:i}=window.freeform.captchas,d=L(r),{lazyLoad:g=!1,version:u="default"}=d,T=g&&!o,c=`${t}-${u}`,l;s.has(c)?l=s.get(c):(l=new Promise((M,H)=>{let w=()=>{P(e).then(M).catch(H)};i.set(c,w)}),s.set(c,l));let E=i.get(c);return T?a.has(n)||(v(n,["input"],E,{once:!0}),a.add(n)):E(),l},m=(e,t)=>t.querySelector(`[data-captcha="${e}"]`),L=e=>({sitekey:e.dataset.sitekey||"",theme:e.dataset.theme||"light",size:e.dataset.size||"normal",version:e.dataset.version,lazyLoad:e.dataset.lazyLoad!==void 0,action:e.dataset.action||"submit",locale:e.dataset.locale});var A="https://challenges.cloudflare.com/turnstile/v0/api.js",y="turnstile",S=(t=>(t.V0="v0",t))(S||{}),U=(e,t)=>{if(!x(e))return Promise.resolve();let o=new URL(A);return o.searchParams.append("render","explicit"),h(o,y,e,t)},x=e=>m(y,e),O=e=>L(e);})();
+(()=>{var E=(e,t,n,o)=>{C("add",{elements:e,type:t,callback:n,options:o})};var C=(e,{type:t,elements:n,callback:o,options:r})=>{let a=Array.isArray(t)?t:[t],s=Array.isArray(n)?n:[n];Array.from(s).forEach(i=>{a.forEach(d=>{e==="add"?i.addEventListener(d,o,r):i.removeEventListener(d,o,r)})})};var p=null;window.freeform||(window.freeform={});var f;(f=window.freeform)!=null&&f.captchas||(window.freeform.captchas={loaders:new Map,listeners:new WeakSet,loaderPromises:new Map});var P=e=>new Promise((t,n)=>{let o=String(e);if(document.getElementById(o))return t();let r=document.createElement("script"),{trustedTypes:a}=window;if(a!=null&&a.createPolicy){if(!p)try{p=a.createPolicy("freeform#captcha-loader",{createScriptURL:s=>s})}catch{}r.src=p?p.createScriptURL(String(e)):String(e)}else r.src=String(e);r.async=!0,r.defer=!0,r.id=o,r.addEventListener("load",()=>t()),r.addEventListener("error",()=>n(new Error(`Error loading script ${e}`))),document.body.appendChild(r)}),u=(e,t,n,o)=>{let r=m(t,n);if(!r)return;let{listeners:a,loaderPromises:s,loaders:i}=window.freeform.captchas,d=L(r),{lazyLoad:g=!1,version:h="default"}=d,T=g&&!o,c=`${t}-${h}`,l;s.has(c)?l=s.get(c):(l=new Promise((M,H)=>{let w=()=>{P(e).then(M).catch(H)};i.set(c,w)}),s.set(c,l));let v=i.get(c);return T?a.has(n)||(E(n,["input"],v,{once:!0}),a.add(n)):v(),l},m=(e,t)=>t.querySelector(`[data-captcha="${e}"]`),L=e=>({sitekey:e.dataset.sitekey||"",theme:e.dataset.theme||"light",size:e.dataset.size||"normal",version:e.dataset.version,lazyLoad:e.dataset.lazyLoad!==void 0,action:e.dataset.action||"submit",locale:e.dataset.locale});var A="https://challenges.cloudflare.com/turnstile/v0/api.js",y="turnstile",S=(t=>(t.V0="v0",t))(S||{}),F=(e,t)=>{if(!x(e))return Promise.resolve();let o=new URL(A);return o.searchParams.append("render","explicit"),u(o,y,e,t)},x=e=>m(y,e),U=e=>L(e);})();
