+++ packages/plugin/src/Resources/js/scripts/front-end/captchas/turnstile/v0.js
@@ -1,1 +1,1 @@
-(()=>{var P={form:{ready:"freeform-ready",reset:"freeform-on-reset",submit:"freeform-on-submit",removeMessages:"freeform-remove-messages",fieldRemoveMessages:"freeform-remove-field-messages",renderSuccess:"freeform-render-success",renderFieldErrors:"freeform-render-field-errors",renderFormErrors:"freeform-render-form-errors",ajaxBeforeSuccess:"freeform-before-ajax-success",ajaxSuccess:"freeform-ajax-success",ajaxError:"freeform-ajax-error",ajaxBeforeSubmit:"freeform-ajax-before-submit",ajaxAfterSubmit:"freeform-ajax-after-submit",afterFailedSubmit:"freeform-after-failed-submit",handleActions:"freeform-handle-actions"},rules:{applied:"freeform-rules-applied"},table:{onAddRow:"freeform-field-table-on-add-row",afterRowAdded:"freeform-field-table-after-row-added",onRemoveRow:"freeform-field-table-on-remove-row",afterRemoveRow:"freeform-field-table-after-remove-row"},dragAndDrop:{renderPreview:"freeform-field-dnd-on-render-preview",renderPreviewRemoveButton:"freeform-field-dnd-on-render-preview-remove-button",renderErrorContainer:"freeform-field-dnd-render-error-container",showGlobalMessage:"freeform-field-dnd-show-global-message",appendErrors:"freeform-field-dnd-append-errors",clearErrors:"freeform-field-dnd-clear-errors",afterErrors:"freeform-field-dnd-after-errors",onChange:"freeform-field-dnd-on-change",onUploadProgress:"freeform-field-dnd-on-upload-progress"},saveAndContinue:{saveFormhandleToken:"freeform-save-form-handle-token"},scripts:{beforeLoad:"freeform-script-before-load",afterLoad:"freeform-script-after-load"},stylesheets:{beforeLoad:"freeform-stylesheet-before-load",afterLoad:"freeform-stylesheet-after-load"}},m=P;var l=(e,t,r,n)=>{R("add",{elements:e,type:t,callback:r,options:n})};var R=(e,{type:t,elements:r,callback:n,options:o})=>{let a=Array.isArray(t)?t:[t],s=Array.isArray(r)?r:[r];Array.from(s).forEach(i=>{a.forEach(c=>{e==="add"?i.addEventListener(c,n,o):i.removeEventListener(c,n,o)})})};var p=null;window.freeform||(window.freeform={});var y;(y=window.freeform)!=null&&y.captchas||(window.freeform.captchas={loaders:new Map,listeners:new WeakSet,loaderPromises:new Map});var F=e=>new Promise((t,r)=>{let n=String(e);if(document.getElementById(n))return t();let o=document.createElement("script"),{trustedTypes:a}=window;if(a!=null&&a.createPolicy){if(!p)try{p=a.createPolicy("freeform#captcha-loader",{createScriptURL:s=>s})}catch{}o.src=p?p.createScriptURL(String(e)):String(e)}else o.src=String(e);o.async=!0,o.defer=!0,o.id=n,o.addEventListener("load",()=>t()),o.addEventListener("error",()=>r(new Error(`Error loading script ${e}`))),document.body.appendChild(o)}),w=(e,t,r,n)=>{let o=E(t,r);if(!o)return;let{listeners:a,loaderPromises:s,loaders:i}=window.freeform.captchas,c=L(o),{lazyLoad:M=!1,version:C="default"}=c,H=M&&!n,d=`${t}-${C}`,f;s.has(d)?f=s.get(d):(f=new Promise((x,S)=>{let A=()=>{F(e).then(x).catch(S)};i.set(d,A)}),s.set(d,f));let g=i.get(d);return H?a.has(r)||(l(r,["input"],g,{once:!0}),a.add(r)):g(),f},E=(e,t)=>t.querySelector(`[data-captcha="${e}"]`),L=e=>({sitekey:e.dataset.sitekey||"",theme:e.dataset.theme||"light",size:e.dataset.size||"normal",version:e.dataset.version,lazyLoad:e.dataset.lazyLoad!==void 0,action:e.dataset.action||"submit",locale:e.dataset.locale});var k="https://challenges.cloudflare.com/turnstile/v0/api.js",T="turnstile";var u=(e,t)=>{if(!v(e))return Promise.resolve();let n=new URL(k);return n.searchParams.append("render","explicit"),w(n,T,e,t)},v=e=>E(T,e),b=e=>L(e);var h=e=>{let t=v(e.form);if(!t)return null;let r=e.form.querySelector(".cl-turnstile");if(r)return r;r=document.createElement("div"),r.classList.add("cl-turnstile"),t.appendChild(r);let{sitekey:n,theme:o,size:a,action:s}=b(t),i=turnstile.render(r,{sitekey:n,theme:o,size:a,action:s});return r.dataset.captchaId=String(i),r},j=e=>new Promise(t=>{let r=0,n=setInterval(()=>{let o=e.querySelector('[name="cf-turnstile-response"]');if(o!=null&&o.value){clearInterval(n),t();return}r+=100,r>=8e3&&(clearInterval(n),t())},100)});document.addEventListener(m.form.submit,e=>{!v(e.form)||e.isBackButtonPressed||e.addCallback(async()=>{try{await u(e.form,!0)}catch{return}h(e),await j(e.form)})});document.addEventListener(m.form.ready,e=>{u(e.form).then(()=>{h(e)}).catch(()=>{})});l(document,[m.form.ajaxAfterSubmit],async e=>{u(e.form,!0).then(()=>{let t=h(e);if(t){let r=t.dataset.captchaId;r&&turnstile.reset(r)}}).catch(()=>{})});})();
+(()=>{var R={form:{ready:"freeform-ready",reset:"freeform-on-reset",submit:"freeform-on-submit",removeMessages:"freeform-remove-messages",fieldRemoveMessages:"freeform-remove-field-messages",renderSuccess:"freeform-render-success",renderFieldErrors:"freeform-render-field-errors",renderFormErrors:"freeform-render-form-errors",ajaxBeforeSuccess:"freeform-before-ajax-success",ajaxSuccess:"freeform-ajax-success",ajaxError:"freeform-ajax-error",ajaxBeforeSubmit:"freeform-ajax-before-submit",ajaxAfterSubmit:"freeform-ajax-after-submit",afterFailedSubmit:"freeform-after-failed-submit",handleActions:"freeform-handle-actions"},rules:{applied:"freeform-rules-applied"},table:{onAddRow:"freeform-field-table-on-add-row",afterRowAdded:"freeform-field-table-after-row-added",onRemoveRow:"freeform-field-table-on-remove-row",afterRemoveRow:"freeform-field-table-after-remove-row"},dragAndDrop:{renderPreview:"freeform-field-dnd-on-render-preview",renderPreviewRemoveButton:"freeform-field-dnd-on-render-preview-remove-button",renderErrorContainer:"freeform-field-dnd-render-error-container",showGlobalMessage:"freeform-field-dnd-show-global-message",appendErrors:"freeform-field-dnd-append-errors",clearErrors:"freeform-field-dnd-clear-errors",afterErrors:"freeform-field-dnd-after-errors",onChange:"freeform-field-dnd-on-change",onUploadProgress:"freeform-field-dnd-on-upload-progress"},saveAndContinue:{saveFormhandleToken:"freeform-save-form-handle-token"},scripts:{beforeLoad:"freeform-script-before-load",afterLoad:"freeform-script-after-load"},stylesheets:{beforeLoad:"freeform-stylesheet-before-load",afterLoad:"freeform-stylesheet-after-load"}},m=R;var l=(e,t,r,n)=>{F("add",{elements:e,type:t,callback:r,options:n})};var F=(e,{type:t,elements:r,callback:n,options:o})=>{let a=Array.isArray(t)?t:[t],s=Array.isArray(r)?r:[r];Array.from(s).forEach(i=>{a.forEach(c=>{e==="add"?i.addEventListener(c,n,o):i.removeEventListener(c,n,o)})})};var p=null;window.freeform||(window.freeform={});var y;(y=window.freeform)!=null&&y.captchas||(window.freeform.captchas={loaders:new Map,listeners:new WeakSet,loaderPromises:new Map});var k=e=>new Promise((t,r)=>{let n=String(e);if(document.getElementById(n))return t();let o=document.createElement("script"),{trustedTypes:a}=window;if(a!=null&&a.createPolicy){if(!p)try{p=a.createPolicy("freeform#captcha-loader",{createScriptURL:s=>s})}catch{}o.src=p?p.createScriptURL(String(e)):String(e)}else o.src=String(e);o.async=!0,o.defer=!0,o.id=n,o.addEventListener("load",()=>t()),o.addEventListener("error",()=>r(new Error(`Error loading script ${e}`))),document.body.appendChild(o)}),w=(e,t,r,n)=>{let o=E(t,r);if(!o)return;let{listeners:a,loaderPromises:s,loaders:i}=window.freeform.captchas,c=L(o),{lazyLoad:C=!1,version:H="default"}=c,x=C&&!n,d=`${t}-${H}`,f;s.has(d)?f=s.get(d):(f=new Promise((S,A)=>{let P=()=>{k(e).then(S).catch(A)};i.set(d,P)}),s.set(d,f));let g=i.get(d);return x?a.has(r)||(l(r,["input"],g,{once:!0}),a.add(r)):g(),f},E=(e,t)=>t.querySelector(`[data-captcha="${e}"]`),L=e=>({sitekey:e.dataset.sitekey||"",theme:e.dataset.theme||"light",size:e.dataset.size||"normal",version:e.dataset.version,lazyLoad:e.dataset.lazyLoad!==void 0,action:e.dataset.action||"submit",locale:e.dataset.locale}),T=(e,t)=>new Promise(r=>{let n=0,o=setInterval(()=>{let a=e.querySelector(`[name="${t}"]`);if(a!=null&&a.value){clearInterval(o),r();return}n+=100,n>=8e3&&(clearInterval(o),r())},100)});var j="https://challenges.cloudflare.com/turnstile/v0/api.js",b="turnstile";var u=(e,t)=>{if(!v(e))return Promise.resolve();let n=new URL(j);return n.searchParams.append("render","explicit"),w(n,b,e,t)},v=e=>E(b,e),M=e=>L(e);var h=e=>{let t=v(e.form);if(!t)return null;let r=e.form.querySelector(".cl-turnstile");if(r)return r;r=document.createElement("div"),r.classList.add("cl-turnstile"),t.appendChild(r);let{sitekey:n,theme:o,size:a,action:s}=M(t),i=turnstile.render(r,{sitekey:n,theme:o,size:a,action:s});return r.dataset.captchaId=String(i),r};document.addEventListener(m.form.submit,e=>{!v(e.form)||e.isBackButtonPressed||e.addCallback(async()=>{try{await u(e.form,!0)}catch{return}h(e),await T(e.form,"cf-turnstile-response")})});document.addEventListener(m.form.ready,e=>{u(e.form).then(()=>{h(e)}).catch(()=>{})});l(document,[m.form.ajaxAfterSubmit],async e=>{u(e.form,!0).then(()=>{let t=h(e);if(t){let r=t.dataset.captchaId;r&&turnstile.reset(r)}}).catch(()=>{})});})();
